refactor: move mock handlers to dedicated single file
