+++ webapp/client/apps/orchestration-cluster-webapp/test/visual/login.test.ts
@@ -9,7 +9,7 @@
 import {test, expect} from '#/pw-modules/test-extend';
 import {HttpResponse} from 'msw';
 import {LoginPage} from '#/pages/Login.page';
-import {mockCurrentUserEndpoint} from '#/shared-test-modules/mockCurrentUser';
+import {mockCurrentUserEndpoint} from '#/shared-test-modules/mock-handlers';
 
 test('should match the login page snapshot', async ({network, page}) => {
 	network.use(mockCurrentUserEndpoint({successResponse: new HttpResponse(null, {status: 401})}));
